backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-migrate==4.0.7
flask-sqlalchemy==3.1.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
requests==2.32.3
